+++ tests/run_unittests.sh
@@ -21,20 +21,6 @@ echo "host=mariadb" >> $ZOOM_DEFAULT_SITE_INI
 echo "user=root" >> $ZOOM_DEFAULT_SITE_INI
 echo "password=root" >> $ZOOM_DEFAULT_SITE_INI
 
-# install google chrome
-wget -q -O - https://dl-ssl.google.com/linux/linux_signing_key.pub | apt-key add -
-sh -c 'echo "deb [arch=amd64] http://dl.google.com/linux/chrome/deb/ stable main" >> /etc/apt/sources.list.d/google-chrome.list'
-apt-get -y update
-apt-get install -y google-chrome-stable
-
-# install chromedriver
-wget -O /tmp/chromedriver.zip $(python3 get_chromedriver_url.py)
-unzip /tmp/chromedriver.zip chromedriver-linux64/chromedriver -d /tmp/
-mv /tmp/chromedriver-linux64/chromedriver /usr/local/bin/chromedriver
-
-# set display port to avoid crash
-export DISPLAY=:99
-
 # get locales not included in base Python image
 apt-get update && apt-get install -y --no-install-recommends \
     locales \
@@ -43,9 +29,6 @@ apt-get update && apt-get install -y --no-install-recommends \
 # set locale
 export LC_ALL=C
 
-# run zoom server
-zoom serve -p 8000 $ZOOM_DEFAULT_INSTANCE &
-
 cat $ZOOM_DEFAULT_SITE_INI
 
-nosetests --with-doctest --with-coverage --cover-package=zoom -vx zoom tests/unittests tests/apptests tests/webtests --exclude-dir=zoom/testing --exclude-dir=zoom/cli
+nosetests --with-doctest --with-coverage --cover-package=zoom -vx zoom tests/unittests --exclude-dir=zoom/testing --exclude-dir=zoom/cli