+++ tests/unittests/test_records.py
@@ -4,8 +4,6 @@
     Test the records module
 """
 
-from decimal import Decimal
-from datetime import date, time, datetime
 import unittest
 
 import zoom
@@ -114,6 +112,8 @@ class TestRecordStore(unittest.TestCase):
 
         person = self.people.get(jane_id)
         person.set(age=29)
+        self.assertEqual(person.age, 29)
+        self.assertEqual(person['age'], 29)
         del person['__store']
         self.assertEqual(
             dict(person),