+++ zoom/__init__.py
@@ -36,7 +36,7 @@ from .page import page, Page
 from .packages import requires
 from .store import store_of
 from .records import table_of
-from .sites import get_site, get_db
+from .sites import get_site, get_db, db
 from .tools import home, redirect_to, load, partial
 from .users import authorize, can, get_user, locate_user
 from .utils import Record