+++ CHANGELOG.md
@@ -6,6 +6,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+## [6.22.0] - 2023-11-20
+- add support for custom MySQL port
+- supress autocomplete by default on date fields
+- add Group methods to get subgroups and supergroups
+- limit path column being inserted to size of column
+- add top level db function
+- add request environment variables to admin environment view
+- add audit logging to Group add_user method
+- add Group remove_user method
+
 ## [6.21.0] - 2023-05-01
 - add quiet logging completion of ajax requests
 - update admin overview to use quiet logging